refactor dds_service
