+++ src/dds_access/dds_data.py
@@ -19,8 +19,7 @@ import copy
 from queue import Queue
 from typing import Dict, List, Optional
 import gc
-
-from dds_access.dds_service import getDataType
+from dds_access.dds_utils import getDataType
 from dds_access.BuiltInObserver import builtin_observer
 from dds_access.dds_qos import qos_match, dds_qos_policy_id
 from utils import singleton, EntityType
